+++ src/_helpers.ts
@@ -128,3 +128,18 @@ export function normalizePackageManifest (data: any): asserts data is normalizeP
     data.version = oVersion.trim()
   }
 }
+
+
+/**
+ * Ignore pattern for `resolved`.
+ * - `^ignore:`- well, just ignore it ... i guess.
+ * - `^file:`  - local dist cannot be shipped and therefore should be ignored.
+ */
+export const npmResolvedIgnoreMatcher = /^(?:ignore|file):/i
+
+/**
+ * VCS pattern for `resolved`.
+ * - `^git:` & `^git+...:` & `*.git`- git
+ * - `^ssh:`                        - probably git
+ */
+export const npmResolvedVcsMatcher = /^git(?:\+[^:]+)?:|\.git(?:#|$)|^ssh:/i